backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
python-multipart
python-dotenv
PyJWT
bcrypt
email-validator
